Enforce a character limit
The two microblogging platforms supported both limit the number of
characters that are allowed in a single notice.

I would do this via autoconf, but I have no idea (and no intention of
learning) how to do it.
This is also permanently on. It needs a flag and a configuration option.