Don't shrink URLs when the message is short
It may happen that, even when posting a URL, the notice in full can be
microblogged without truncation, and other users get a link of blind
redirection.

Create a configuration option, and reuse existing flags, to specify when
to "always" use a URL-shrinking service (default), and when to shrink
only when the post would be too long.

The configuration option is not entirely backwards-compatible, but to
make it backwards-compatible, the configuration would be
forwards-incompatible.